alsa-utils: Add patch to fix record many files when using arecord command
When using arecord, it will generate many files. It's necessary to add the patch.

root@imx6sxsabresd:~# arecord -Dhw:0 -d 10 -f S16_LE -c 2 -r 11025 a.wav
root@imx6sxsabresd:~# ls
a-01.wav  a-06.wav  a-11.wav  a-16.wav  a-21.wav  a-26.wav  a-31.wav  a-36.wav
a-02.wav  a-07.wav  a-12.wav  a-17.wav  a-22.wav  a-27.wav  a-32.wav  a-37.wav
a-03.wav  a-08.wav  a-13.wav  a-18.wav  a-23.wav  a-28.wav  a-33.wav  a-38.wav

Signed-off-by: Zidan Wang <b50113@freescale.com>