mesa: Fix 6UL Wayland build [YOCIMX-349]
The mesa recipe makes a textual change to eglplatform.h, but
the source text was changed from '#ifdef' to '#if defined'.
Our build is using the old version of the header, with the new
version of the recipe, so the header was not being updated,
which caused other recipes to fail due to undefined X11 headers.