Addition of packagegroups for networking. The core packagegroup file contains basic networking tools. If the openembedded network layer is available, tools from that layer will be used. It also contains a server packagegroup that can be used if the network layer is present.
The extended packagegroup contains additional networking tools that
may be desired in a larger, more complex image. Tools from the networking
layer will be included if it is available.

These packagegroups can help to build image recipes without having to
be an expert on network tools and make the recipe more readable.

Signed-off-by: Ann Thornton <ra43240@freescale.com>