sysvinit: update rc file to fix MX6QP SDB eMMC stress reboot fail
The reason for the blocking is that: there is the race condition that when rc
process try to write the PIPE but find there is no consumer(reader) for it,
then it will wait for the consumer and schedule out.
The background psplash process will be terminated silently when X up/down
and thus will cause this race condition.

Signed-off-by: Yuqing Zhu <b54851@freescale.com>