kernel-module-imx-gpu-viv: Add kernel dependency as workaround [YOCIMX-530]
The kernel dependency is supposed to be given to us by inheriting
module.bbclass, but it doesn't seem to be working. We're seeing two
build breaks: 1) at compile time when do_rm_work is done on
linux-imx before compiling kernel-module-imx-gpu-viv, and 2) at
do_rootfs when kernel-module-galcore fails to rebuild after an update
to linux-imx.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>