pulseaudio: use the imx6ull specific daemon.conf as the common one for all platforms [YOCIMX-831]
The new daemon.conf is used to increase the rlimit-rttime for imx6ull.
The performance of imx6ull seems not good enough, it cause the pulseaudio exit
when cpu loading very high. The rlimit-rttime specifies a limit on the amount
of CPU time that a process scheduled under a real-time scheduling policy may
consume without making a blocking system call.  If the pulseaudio cost more
time which exceed the rlimit-rttime, it will be killed by kernel.
So here we increase the rlimit-rttime to workaround this issue.

And it will have no adverse impact on other platform.

So unify the daemon.conf with one new version for all platforms.

Signed-off-by: Yuqing Zhu <carol.zhu@nxp.com>