Revert "qtbase: Fix order so that 3d is last and does not pick up 2d overrides as default" [YOCIMX-1021]
This reverts commit b754cff5b8cad189b29abebd7b9392041dac098b.

Reverting the commit because it has no effect and it skews the
file relative to its upstream counterpart in meta-freescale.

Determining which overridden variable value to use is not based on
the relative order of the variables in the recipe. It is instead
based on the relative order of the overrides in the machine
configuration, with higher priority given to overrides that are
further right on the list.

Example:

    A          = ""
    A_mx6      = "mx6"
    A_imxgpu3d = "3d"
    A_imxgpu2d = "2d" // Not best practice [1]

1. For an i.MX 6Quad board, the machine overrides looks like this:

    mx6:mx6q:imxipu:imxvpu:imxgpu2d:imxgpu3d

The value of A is "3d".

2. For an i.MX 6SoloLite board, the machine overrides looks like this:

    imxpxp:imxgpu2d:imxepdc

The value of A is "2d".

[1] The order of the variables is chosen to illustrate the fact
    that the order is not important. Best practice suggests that
    the variables should be ordered in the file to match the
    machine override priority:

    A          = ""
    A_mx6      = "mx6"
    A_imxgpu2d = "2d"
    A_imxgpu3d = "3d"