fsl-setup-release.sh: Fix workspace problems [YOCIMX-1102]
Unfortunately Yocto does not provide a mechanism that allows us
to override a machine configuration file or a bbclass file in
the meta-freescale layer. To overcome this, we have been copying
all of our machine and bbclass files to meta-freescale.

However, this solution is problematic whenever someone pushes a change
to one of these files. Once the workspace is updated via 'repo sync',
it is put into an incorrect state since it will continue to use the
file that was last copied by the script, which is now stale. What's worse,
the problem that manifests due to the error will have something to do with
the change that was made to the file, and an un-necessary debug cycle
will ensue.

This problem is fixed by deleting the upstream file instead of overwriting
it. With the upstream file removed, the updated file in our layers is used,
including after every repo sync.

This solution has one potential problem. If a modified file is removed
from our internal or release layer, then a repo sync will result in
a workspace that is missing the file. Unlike the previous case, the
problem that manifests due to this error will be quite direct, and it will
be simple to restore the file to meta-freescale.

The fix also has a side benefit of only deleting upstream files where a
conflict exists. The previous solution did copy all machine files
even though all but a few were needed to reside in meta-freescale.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>