imx-lib: Update to commit 67083e1
Update the last commit info, it should follow the description pattern,
or it will cause some build errors.

eg:
The comparsion function will say "two commits are not equal"
because of the patch description.
then do: git commit -s -m
but, nothing to commit
in the end, it will break.

Signed-off-by: Yuqing Zhu <carol.zhu@nxp.com>