connman: Fix network interface when booting over NFS [YOCIMX-1193]
The network interface was not working after booting over NFS due to
interference by connman. We solved this same problem in the past by
configuring connman in sysvinit to ignore any such interface. Part
of that solution was adapted to our current systemd startup, but the
part for detecting the NFS mount was omitted erroneously. This commit
restores the detection logic and aligns the systemd and sysvinit
solutions.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>