conf/distro/include/fsl-imx-base.inc: Possible fix for OE-core config sanity build break
There is a common build break on our build servers:

ERROR:  OE-core's config sanity checker detected a potential misconfiguration.
    Either fix the cause of this error or at your own risk disable the checker (see sanity.conf).
    Following is the list of potential problems / advisories:

    Fetcher failure for URL: 'https://eula-downloads.yoctoproject.org/index.php'. URL https://eula-downloads.yoctoproject.org/index.php doesn't work. Please ensure your network is configured correctly.

The URLs used for this test are specified in our distro config. The current values
were copied when some part of meta-yocto/conf/distro/poky.conf was copied into our
distro configuration.

The source of the copy has since been updated, replacing the original URLs
with a more reliable (and faster) choice.

http://git.yoctoproject.org/cgit.cgi/poky/commit/meta-yocto/conf/distro/poky.conf?h=jethro&id=16d64def97154d76fc1134ecd6cf9d704621e689

At the least, it's good to be in sync with the up-layer value. At the best,
hopefully we will stop having the frequent build breaks.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>