imx-gpu-viv: Fix internal recipe build break [YOCIMX-1763]
New logic was added to the include file for conditionally installing
gbm.pc only for 3d parts. This logic was failing when building the
internal recipe since the internal recipe has a different install
design. This is fixed by moving the new logic directly into the
do_install function, which is fully overridden by the internal recipe.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>