imx-test: Simplify and clarify dependency on VPU
This change should have no logical change.
I was surprised to see that virtual/imxvpu was being
used for parts without VPU. I thought to fix it, but
after digging in the git history found that this was
a feature, not a bug.

https://bitbucket.sw.nxp.com/projects/IMX/repos/meta-fsl-bsp-release/commits/d03342173b6eec922d015230c8b28fc579034ff1#imx/meta-bsp/recipes-bsp/imx-vpu/imx-vpu_5.4.32.bb

I've updated the DEPENDS logic and included a comment
to make this clear.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>