libproxy: Fix multi-machine build breaks
When building for imx7ulpevk, a build break was occurring
due to a missing glibc library crt1.o.

The library was missing because the do_configure task was
using an old imx6ul7d sysroot instead of the correct
imx7ulpevk sysroot.

The wrong sysroot was being used by the build system
because it determined that the run.do_configure task
from a prior 6ul7d build was equivalent.

Setting PACKAGE_ARCH to be machine-specific fixes the
re-use issue.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>