alsa-lib: Refactor bbappend file [YOCIMX-2202]
The patches are compatible with current i.MX platforms.
But we couldn't guarantee the legacy ones,
So add back patch applying filter, keeping align
with the patch applying method in meta-freescale layer.

Remove "0004-add-conf-for-multichannel-support-in-imx.patch"
as it's duplicate with "0001-add-conf-for-multichannel-support-in-imx.patch"
under meta-freescale layer.

"0004-pcm-Don-t-store-the-state-for-SND_PCM_STATE_SUSPENDE.patch"
from meta-freescale layer is ok for mx6/mx7,
but not used in i.MX8 previous release, so remove it to keep the same as before.
Will add it back if audio team make request.

Signed-off-by: Yuqing Zhu <carol.zhu@nxp.com>