imx-gpu-sdk: Fix --UseExtensions configuration [MGS-3199]
The --UseExtensions option was not working correctly and was
in fact enabling all extensions no matter the setting. That is
now fixed in the source.

Given that, the settings specified in the recipe were incorrect
anyway. Now we enable all extensions in the default case with
--UseExtensions [*], and we specify just a single extension
valid for the 8MQ case.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>