wayland-ivi-extension_git.bb: Add recipe
Add git version of wayland-ivi-extension to include changes which are
not released yet. We need this for compatibility with weston 4, since we
are bumping our version of weston.
It's mainly used in our GENIVI and AGL builds; adding it here instead of
maintaining the same recipe in two separate layers.

Signed-off-by: Mihai Lindner <mihai.lindner@nxp.com>