weston: Fix patch fuzz warning
This fixes the following warning:

WARNING: weston-5.0.0.imx-r0 do_patch:
Some of the context lines in patches were ignored. This can lead to incorrectly applied patches.
The context lines in the patches can be updated with devtool:

    devtool modify <recipe>
    devtool finish --force-patch-refresh <recipe> <layer_path>

Then the updated patches and the source tree (in devtool's workspace)
should be reviewed to make sure the patches apply in the correct place
and don't introduce duplicate lines (which can, and does happen
when some of the context is ignored). Further information:
http://lists.openembedded.org/pipermail/openembedded-core/2018-March/148675.html
https://bugzilla.yoctoproject.org/show_bug.cgi?id=10450
Details:
Applying patch 0003-weston-touch-calibrator-Advertise-the-touchscreen-ca.patch
patching file weston.ini.in
Hunk #1 succeeded at 3 with fuzz 2 (offset 1 line).

Now at patch 0003-weston-touch-calibrator-Advertise-the-touchscreen-ca.patch

Signed-off-by: Neena Busireddy <neenareddy.busireddy@nxp.com>