fsl-dynamic-packagearch.bbclass: Add a new package_arch for mx8 packages [YOCIMX-3004]
- This mechanism allows to reuse common GPU packages among all mx8 SoCs.
- Some of the GPU packages are no longer SOC specific and thus packages that are dependent
on those packages will not get rebuilt. This reduces the build time.

Example:

Chromium was rebuilt for all the SoC's as virtual/egl was Soc specific and chromium was
dependent on it. With this mechanism, chromium is built only once since virtual/egl will be
using same package_arch amoung all Soc's

Signed-off-by: Neena Busireddy <neenareddy.busireddy@nxp.com>