linux-libc-headers: Remove i.MX and SDK-specific version
The idea of using custom headers in the SDK only was meant to
make it easier to deal with custom kernel headers in the SDK
context. Now the thinking is we don't need to make it easier,
and we will instead add any custom kernel headers through
linux-imx-headers.

By doing this, the custom headers will be perhaps more visible,
which may help accelerate the process of getting them
upstreamed and removed from linux-imx-headers.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>