fsl-imx-base.inc: Fix warning for Ubuntu 18.04 [YOCIMX-3479]
The variable SANITY_TESTED_DISTROS was copied from poky some time
ago and is now stale. It's not clear why it was copied. Remove it
to simplify matters and fix the following warning:

WARNING: Host distribution "ubuntu-18.04" has not been validated with this version of the build system; you may possibly experience unexpected failures. It is recommended that you use a tested distribution.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>