rng-tools/rngd.service: Fix reboot 90 seconds delay [YOCIMX-3738]
There's a delay when rebooting because the rngd service doesn't get
a stop signal and ends up being killed after 90s of wait time.
This is fixed in poky, master branch, but not yet picked up by warrior,
maybe never will. We override rngd.service file, until the fix gets to
warrior or we switch to zeus, currently in development / master.

Signed-off-by: Mihai Lindner <mihai.lindner@nxp.com>