MACHINE.conf: Replace machine config override model [YOCIMX-3918]
The machine configuration override model was based on maintaining
a copy of the upstream meta-freescale machine configuration files.
The argument against this model is no different than the argument
for using bbappend for a recipe instead of BBMASK and a copy of
the recipe. The copied logic gets out of date easily and is hard
to synchronize with meta-freescale.

Use a simpler model that overrides existing machine configuration
in layer.conf. With this model, only changes to the existing
configuration need to be maintained.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>