imx-uuc: Fix rootnfs incompatibility with connman [MLK-22828]
After using uuu to set up a board, the image was unable to mount the
rootfs over nfs. It seems that connman was unable to detect the rootnfs
connection over usb interface and was overwriting the configuration.
The problem is fixed in uuu by adapting the bootargs to the format
that connman is expecting.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>