weston-init: Set XDG_RUNTIME_DIR if it's not already [YOCIMX-4018]
Running weston over ssh was not working:

error: XDG_RUNTIME_DIR not set in the environment

XDG_RUNTIME_DIR is normally set by pam with systemd, but not in
the case of ssh.

Work around the problem by setting XDG_RUNTIME_DIR with the same
logic implemented for sysvinit in weston-start.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>