weston-init: Fix IVI desktop [LF-390]
The upstream weston.ini was recently overridden for YOCIMX-4086,
but the override is incorrectly applying to the IVI case as well,
causing the IVI image to come up with the standard weston desktop
instead of the IVI desktop.

To fix the IVI desktop, the non-IVI override must itself be
overridden to use the original IVI version of weston.ini.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>