base-files: Set serial console window size properly [YOCIMX-4200]
The serial console behaves as if it is 80x24 no matter the actual window
size. The base version of profile fixed this five years ago with a resize
command, but we overrode the base version and removed the resize in order
to avoid a less important issue in YOCIMX-1180.

Remove our override and just use the default profile. Note that I was
not able to reproduce YOCIMX-1180.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>