layer: Adapt to improved EULA handling model [YOCIMX-4322, YOCIMX-4323]
The EULA handling class fsl-eula-unpack in meta-freescale did not
allow for updates to the EULA file itself. Until now we've resorted
to replacing the EULA file and the class file, a messy solution.

The class fsl-eula-unpack is now improved to properly handle EULA
updates. Adapt the layer to the new EULA update model:
- imx-setup-release.sh: Don't copy EULA to meta-freescale
- layer.conf:
  - Set the current EULA to LA_OPT_NXP_Software_License v11 February 2020
  - Note the v10 EULA since meta-freescale is currently at v9
- fsl-eula-unpack.bbclass: Remove our copy of this file

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>