linux-imx: Add dependence on do_patch for task copy_defconfig
When using other metalayer to bbappend the linux-imx recipe, if the
bbappend includes a patch on defconfig, it may or may not be effective
during the do_configure task, depending on how the task chain is
executed.

This patch prevents that behavior and force that the copy_defconfig
happens always after any do_patch (from any bbappend) is executed.

While at this, remove the dependency on unpack as it is obsolete now.

Signed-off-by: Daiane Angolini <daiane.angolini@nxp.com>