imx-image-full: Move fully into qt5 layer [YOCIMX-4431]
The SDK for test-internal-qt5 was missing the Qt5 tools.
The image was including imx-image-full, but including
a recipe does not include the recipe's bbappends, so the
imx-image-full bbappend that added Qt to the recipe
was not doing anything for test-internal-qt5, which is
why the SDK was incomplete.

Since imx-image-full is in fact meant to include Qt5,
the recipe can be moved directly into the qt5 dynamic
layer, simplifying the image recipe design and fixing
the problem with the SDK.

This change also removes redundant lines in
fsl-image-qt5-validation-imx recipe.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>