imx-gpu-viv: Remove unneeded libraries from install [YOCIMX-4495]
The upstream install doesn't know about the newly versioned
libraries and causes a QA issue for the libvulkan libs that
are not for the current backend.

ERROR: imx-gpu-viv-1_6.4.0.p2.4-aarch64-r0 do_package: QA Issue: imx-gpu-viv: Files/directories were installed but not shipped in any package:
	  /usr/lib/libvulkan-wl.so.1.1.3
	  /usr/lib/libvulkan-fb.so.1.1.3

The files are not needed in the install, so simply remove them.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>