imx-gpu-apitrace: Remove waffle for 6 and 7 except for wayland [YOCIMX-4632]
The following runtime error is observed when apitrace is
built with waffle on 6 and 7 with framebuffer:

```
error: waffle was built without support for WAFFLE_PLATFORM_WAYLAND. WAFFLE_ERROR_BUILT_WITHOUT_SUPPORT(19)
```

Since waffle is not supported for 6 and 7 on framebuffer,
fix the runtime error by removing it. Note that waffle works
fine on Wayland, so waffle remains enabled for 6 and 7 on
Wayland.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>