imx-gpu-viv: Fix rootfs conflict with libvulkan-dev [YOCIMX-4657]
The rootfs stage is failing with the following error:

```
Some packages could not be installed. This may mean that you have
requested an impossible situation or if you are using the unstable
distribution that some required packages have not yet been created
or been moved out of Incoming.
The following information may help to resolve the situation:

The following packages have unmet dependencies:
 libvulkan-imx-dev : Conflicts: libvulkan-dev
                     Recommends: libglslc-imx-dev but it is not installable
                     Recommends: libvsc-imx-dev but it is not installable
E: Unable to correct problems, you have held broken packages.
```

It seems this is due to logic upstream that disables debian
renaming. This logic was originally added due to a conflict with
a recipe that is no longer being used [1]. Restoring the overrides
for libvulkan-imx to the default state fixes the build break.

[1] https://bitbucket.sw.nxp.com/projects/IMX/repos/meta-imx/commits/0e0acfba04e9ada9ef587e759af68c722d19ce27

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>