libvulkan-imx: Add package via machine configuration
The vulkan drivers are loaded dynamically and so must be manually added
to the image. This is done so far from a couple of vulkan recipes, but
it makes those recipes i.MX specific.

A better approach is to add the drivers from the machine configuration
based on the vulkan distro feature.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>