linux-imx: Cleanup do_copy_defconfig [YOCIMX-5212]
Since do_symlink_kernsrc is before do_patch is before do_kernel_configme,
both do_patch and do_symlink_kernsrc dependencies are redundant and can
be dropped.

Adding do_kernel_configme as a dependency helps to expose an important
part of this implementation that is not clear, that the task
do_copy_defconfig does effectively disable the kernel config fragments
provided by do_kernel_configme. A clearer representation of this would
be to disable do_kernel_configme. However, this leads to a failure in
do_kernel_configcheck as it checks for certain output from
do_kernel_configme. So the best we can do here is an added comment.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>