armnn: Add pyarmnn examples, unit tests, whl, and remove mobilenet v1 model [YOCIMX-5262]
Missing pyarmnn examples and a python wheel (for user to install in python virtual environments) are being added to the main package. Pyarmnn unit tests (without setup.py test, which is supposed to run only on host) are being added to the new -tests package. Mobilenet v1 model with labels is being removed (unknown reason why it was there in the first place - the user should copy this on his own). All binaries will be placed in <bindir>/armnn-<version>.

Signed-off-by: Pavel Macenauer <pavel.macenauer@nxp.com>