isp-imx: Normalize the recipe source folder [YOCIMX-5270]
The source folder of the recipe is adjusted to the appshell sub-folder to facilitate
building of that folder. A side effect of this is that the reference to the license
file at the root must also be adjusted. This second adjustment makes automatic
updates of the license complex, and in fact at the moment broken.

Restore the source folder to the default and use OECMAKE_SOURCEPATH to direct the
build to the appshell sub-folder.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>