vulkan-loader: Add libvulkan.so to main package [YOCIMX-5293]
Vulkan apps are failing after the upgrade to 1.2.135.0 because
libvulkan.so is missing. The local recipe for the previous version
contained a fix for this, but it was overlooked when the recipe was
dropped for the newer upstream recipe.

Fix the regression by adding the fix to the local bbappend.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>