imx*.conf: Add old machines for compatibility
Add machine configuration files for previous machine names to preserve
compatibility. This includes all i.MX 8 machines and 6ULZ 14x14 EVK.

Include the new machine file name in the list of overrides so variables
with the overrides are properly found.

Some of the new machine configuration uses the MACHINE variable directly
in the KERNEL_DEVICETREE_BASENAME variable. For these we must set
KERNEL_DEVICETREE_BASENAME to the new machine name.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>