opencv: Rename patch folders for clarity
The patch folders are named `files` for the upstream patch files and
`opencv` for the local. It's clearer to use `opencv` for the upstream
and `opencv-4.5.1.imx` for local.

Note that the bitbake system gives a versioned folder name higher
precedence, which is exactly what is needed whenever an upstream patch
file must be overridden locally.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>