layer.conf: Allow to override variable removals
Currently there is no way to easily add jailhouse to MACHINE_FEATURES
because it is being removed directly _remove operation. The same applies
for SOC_DEFAULT_IMAGE_FSTYPES. See [1].

[1] https://wiki.yoctoproject.org/wiki/Technical_FAQ#How_do_I_remove_a_value_from_a_list_variable.3F

Move the removed items to a variable so that what is being removed can
be changed easily. For example, to restore jailhouse, the following
can be added to local.conf:

MACHINE_FEATURES_IMX_REMOVALS_remove = "jailhouse"

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>