imx-image-*: Fix boot to weston [YOCIMX-5350]
The images imx-image-core and imx-image-multimedia plus their derivatives
no longer boot into weston, while core-image-weston works fine. This is
because systemd is configured for weston boot only if weston is in
IMAGE_FEATURES.

https://github.com/openembedded/openembedded-core/commit/6b6287b7e5190d1af2e3aefddb2f72a8d0d79968

Add the weston feature to the image recipes. Also, remove the addition
of weston packages since the feature now takes care of that.

Also, for imx-image-core, add weston-xwayland since it is missing.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>