weston-init: Cleanup manual XDG configuration [YOCIMX-5350]
The manual configuration for XDG done by the weston profile configuration
file is no longer relevant now that PAM is setting up the user XDG session.

However, after removing the manual configuration, the app weston-simple-egl
fails when run from ssh:

```
root@imx8mpevk:~# weston-simple-egl
has EGL_EXT_buffer_age and EGL_EXT_swap_buffers_with_damage
Segmentation fault
```

This is because XDG_RUNTIME_DIR is not set in the user's environment
for the ssh session. Whether this is a bug or by design is not clear.
For now choose to add a workaround of setting the variable manually.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>