packagegroup-imx-ml: Limit package inclusion to i.MX 8 with GPU [YOCIMX-5453]
Use ML_PKGS variable to include main packages only for i.MX 8 with GPU.
Add a new variable for eiq-apps since it has different limits.

This change makes the packagegroup empty if not an i.MX 8 with GPU, so
an image recipe that includes it does not have to deal with adding such
conditional logic itself.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>