imx-boot: allow for additional/different u-boot device-tree [YOCIMX-5353]
Align with upstream f607eb8734:

With the current call to imx-mkimage and SOC_TARGET = "iMX8M" the u-boot
device tree file name is set to $(PLAT)-evk.dtb in iMX8M/soc.mak which
happens to match ${UBOOT_DTB_NAME} if building for an NXP machine.

Setting the dtbs environment variable to ${UBOOT_DTB_NAME} makes the build
succeed also for other machines.

In the hardcoded version $(dtbs) is an intermediate file and gets delete
as the last step of make. Thus for several make targets we have to redo
the copy step before each make invocation or the second one will fail as
the device tree file is then missing.

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>
Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>