tim-vx: Drop unit tests and proxies [AIR-4841]
The unit tests trigger a manual download of googletest in the configure
task. On certain machines this download fails due to firewall issues.
The network proxies were added as a workaround.

The manual download is the root problem. The right way to fix this is
to add googletest to the dependencies so the package can use googletest
from the Yocto sysroot. This was tried but unfortunately it does not
work, and the package continues to download the package manually.

However, the unit tests are not even needed for the Yocto build.
Disable them and ignore the unit test proxy issue.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>