packagegroup-imx-security: Don't include smw-tests without optee [YOCIMX-5494]
The build for i.MX 8ULP EVK is failing:

```
ERROR: Nothing PROVIDES 'optee-os' (but /opt2/hardknott/sources/meta-imx/meta-bsp/recipes-security/smw/smw_git.bb DEPENDS on or otherwise requires it)
optee-os was skipped: missing required machine feature 'optee' (not in MACHINE_FEATURES)
```

Support for optee is not yet available for 8ULP, and the machine correctly
does not have optee listed in its machine features. The failure is because
the security packagegroup includes smw-tests, which depends on optee.

Fix the error by including smw-tests only if optee is enabled.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>