imx6ul7d: Add missing firmware-imx-sdma-imx7d
The imx6ul7d machine is used to test both 6ul and 7d boards, but only
firmware-imx-sdma-imx6q is included in the rootfs, and the firmware
fails to load from the sysfs fallback on 7d:

```
[   23.110273] imx-sdma 30bd0000.sdma: Direct firmware load for imx/sdma/sdma-imx7d.bin failed with error -2
[   23.239860] imx-sdma 30bd0000.sdma: Falling back to sysfs fallback for: imx/sdma/sdma-imx7d.bin
[   23.292397] imx-sdma 30bd0000.sdma: external firmware not found, using ROM firmware
```

This was working before because the imx6q and imx7d firmware were
packaged and included in the same generic firmware-imx-sdma package.
Now the imx6q and imx7d firmware are in separate packages, aligning
with the pre i.MX 6 SDMA firmware packages.

The mechanism for adding the appropriate SDMA firmware is the override,
but since the imx6ul7d 'combined' machine does not include mx7 in its
MACHINEOVERRIDES, the imx7d firmware is missing. Add a new override
specific to imx6ul7d machine.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>