optee-os: use tee-raw.bin binary [YOCIMX-5659]
During the optee-os compilation, the gen_tee_bin.py script generates
a tee-raw.bin that is more space efficient and could replace
the tee.bin binary manually generated in the optee-os recipe
by objcpy.

This patch greatly reduces the size of tee.bin. On 64 bits platforms, we
observe a size descrease of ~68%, and on 32 bits platforms, a size
decrease of ~45%.

Signed-off-by: Clement Faure <clement.faure@nxp.com>