imx-mkimage: Update imx-mkimage_git.inc to commit 1527fb9
[YOCIMX-5672] Pick up the fixing - iMX8M: fix FIT DT node unit address

The DT spec requires a unit-address of a node name to match the "reg"
property in that node. The latest dtc versions generates a warning on
nodes that specify an unit-address via the '@' but don't have a 'reg'
property.
Also u-boot mkimage enabled with CONFIG_FIT_SIGNATURE will check the
'@' and returns fdt check error.

So update the FIT generation script to use '-' to replace '@'

Signed-off-by: Jun Zhu <junzhu@nxp.com>