imx-boot/u-boot: Add boot image size to SD card image [YOCIMX-3910]
Append a tag to the boot image used in the SD card image. The tag
contains the size of the boot image so UUU can easily find the end of
the boot part in the SD card image.

To implement the feature, an additional tagged image is deployed from
the u-boot-imx and imx-boot recipes, called u-boot-tagged.imx and
imx-boot-tagged respectively.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>