imx-gpu-viv: Fix qtbase break on Wayland [MGS-6424]
The qtbase build on Wayland is broken. The package sets EGL_API_FB.
This setting used to have a lower precedence in eglplatform.h and so
was not a problem. Now that EGL_API_FB is given top precedence, qualify
it so it is not used if Wayland is also set.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>