imx-gpu-viv: Add glu.pc to fix mesa-demos [YOCIMX-5790]
The mesa-demos build is failing with undefined references to glew like
the following:

```
.../mesa-demos-8.4.0/src/fp/tri-param.c:66: undefined reference to `__glewProgramLocalParameter4fARB'
.../mesa-demos-8.4.0/src/fp/tri-param.c:107: undefined reference to `glewInit'
```

Adding glu.pc for the newly added libglu package fixes the problem. Note
that this is counterintuitive to the fact that mesa-demos is configured
without glu support. Perhaps the build configuration logic for mesa-demos
doesn't handle well the case where glu.pc is missing, but the problem is
undetected since glu.pc is not normally missing.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>