linux-imx: Rework kernel source override mechanism
The current implementation that overrides the kernel URL adds
nobranch=1 after branch=next and does not work. The nobranch
option should replace the branch option.

Move the kernel branch specification into the variable KERNEL_SRC.
This makes it easier to then replace the branch option with nobranch=1.

Note that nobranch=1 may not be reliable by itself and may require
a cleanall.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>