mesa-demos: Add GLU for i.MX GPU [LF-5339]
Support for GLU is removed from i.MX GPU drivers. This results in a build
break due to a dependence on GLU. Since GLU is not even configured, this
uncovers a problem with the package build.

Avoid the package build configuration problem by simply configuring to
include GLU, and resolve the GLU dependency with the mesa libglu package.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>