weston: Fix SRC_URI override
The meta-freescale override of SRC_URI for the i.MX fork is more
precise now, taking all of the inherited SRC_URI and changing only what
is needed. The new design breaks the override here.

Fix the problem by using the same design here. Note that this is
complicated because the override here does not change anything if
WESTON_SRC is not changed, and the logic must take this into account.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>