meta-sdk: Fix weston by aligning DISTROOVERRIDES
The workaround to provide XDG_RUNTIME_DIR to ssh is upstreamed to
meta-freescale-distro, but it is qualified to work only if
DISTROOVERRIDES contains fsl or fslc. Our layer is not aligned
with this and ends up with the default setting of DISTROOVERRIDES
equal to the distro name.

Align DISTROOVERRIDES for our meta-imx distros so the qualifier
in meta-freescale-distro will work as expected.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>