vulkan-validationlayers: Upgrade 1.3.204.0 -> 1.3.204.1 [YOCIMX-5998]
The upgrade to 1.3.204.0 has a build break for missing library
/usr/lib/librt.so:

```
ERROR: vulkan-validationlayers-1.3.204.0-r0 do_compile: ExecutionError('/opt2/kirkstone/xwayland/tmp/work/cortexa53-crypto-poky-linux/vulkan-validationlayers/1.3.204.0-r0/temp/run.do_compile.3751972', 1, None, None)
ERROR: Logfile of failure stored in: /opt2/kirkstone/xwayland/tmp/work/cortexa53-crypto-poky-linux/vulkan-validationlayers/1.3.204.0-r0/temp/log.do_compile.3751972
Log data follows:
| DEBUG: Executing shell function do_compile
| NOTE: VERBOSE=1 cmake --build /opt2/kirkstone/xwayland/tmp/work/cortexa53-crypto-poky-linux/vulkan-validationlayers/1.3.204.0-r0/build --target all --
| ninja: error: '/usr/lib/librt.so', needed by 'layers/libVkLayer_khronos_validation.so', missing and no known rule to make it
| WARNING: exit code 1 from a shell command.
```

The library is in the recipe-sysroot folder, but the build is unable to
find it. The root cause of the problem is not known, so a workaround is
added at the end of the configure task to update the path in build.ninja
to point directly to the library in the recipe-sysroot folder.

This leads to two more missing libaries, one for -lrt and one for
-lSPIRV-Tools-static.

Upgrading to 1.3.204.1 fixes both of these problems.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>