weston-init: Stop running weston as root [YOCIMX-6060]
A little over a year ago weston was switched from running as root to
running as weston. Running as weston didn't work properly on i.MX, so
for i.MX the change was overridden.

It is time to re-assess the situation, so remove the override and run
weston as weston.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>