weston: Clean up bbappend
The weston recipe in the layer is confusing since it contains both a
main recipe .bb and a .bbappend. On top of that, the main recipe is
itself a copy of the OE-core recipe plus i.MX overrides. The bbappend
contains regular overrides plus a SRC_URI override needed to override
the main recipe SRC_URI.

Clean up the bbappend by moving the regular overrides to the main
recipe and by clearly commenting the purpose of the SRC_URI overrides.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>