deepview-rt: Fix missing OpenVX library [YOCIMX-6234]
The library deepview-rt-openvx.so is missing from the rootfs. This is
because the package config for OpenVX is removed, which causes logic
in the install to always remove the library deepview-rt-openvx.so from
the install directory.

Move the library into a separate package for OpenVX and include it with
an SOC-specific runtime dependency.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>