imx-base-extend.inc: Fix preferred provider settings for GPU
First, the preferred provider settings for opencl-headers and
opencl-icd-loader are upstream, so just drop them.

Second, the preferred provider settings for GPU belong in machine
config, not in distro config, so move the provider setting for
virtual/libgles3:imxgpu3d to imx-base-extend.inc.

Third, the preferred provider setting for virtual/opencl-icd is
missing, so add it.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>