deepview-rt: Fix non-GPU runtime dependency [YOCIMX-6234]
Non-accelerated builds fail due to a dependency on libopenvx-imx:

```
ERROR: Nothing RPROVIDES 'libopenvx-imx' (but sources/meta-imx/meta-ml/recipes-libraries/deepview-rt/deepview-rt_2.4.42-aarch64.bb RDEPENDS on or otherwise requires it)
imx-gpu-viv RPROVIDES libopenvx-imx but was skipped: incompatible with machine imx8mpul-lpddr4-evk (not in COMPATIBLE_MACHINE)
imx-gpu-viv RPROVIDES libopenvx-imx but was skipped: The kernel of machine needs to have Vivante kernel driver support for this recipe to be used.
imx-gpu-viv RPROVIDES libopenvx-imx but was skipped: The kernel of machine needs to have Vivante kernel driver support for this recipe to be used.
imx-gpu-viv RPROVIDES libopenvx-imx but was skipped: incompatible with machine imx8mpul-lpddr4-evk (not in COMPATIBLE_MACHINE)
NOTE: Runtime target 'libopenvx-imx' is unbuildable, removing...
Missing or unbuildable dependency chain was: ['libopenvx-imx']
ERROR: Required build target 'deepview-rt' has no buildable providers.
Missing or unbuildable dependency chain was: ['deepview-rt', 'libopenvx-imx']
```

Unfortunately bitbake fails on this even though it is deepview-rt-openvx
that is marked with the runtime dependency.

Avoid the problem by moving the runtime dependency to the main package.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>