deepview-rt: Fix non-acclerated packaging [YOCIMX-6234]
Including the accelerated library deepview-rt-openvx.so in a non-
accelerated package leads to the following build failure:

```
ERROR: deepview-rt-2.4.42-aarch64-r0 do_package_qa: QA Issue: /usr/lib/deepview-rt-openvx.so contained in package deepview-rt-openvx requires libOpenVX.so.1()(64bit), but no providers found in RDEPENDS:deepview-rt-openvx? [file-rdeps]
```

It makes no sense to include the library in this context, so use a
package config to exclude it.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>