weston: Use imxgpu flag for i.MX 93 [YOCIMX-6311]
The build for i.MX 93 is failing:

```
| /.../gcc/aarch64-poky-linux/11.3.0/ld: libweston/backend-drm/drm-backend.so.p/kms.c.o: in function `drm_pending_state_apply':
| /.../libweston/backend-drm/kms.c:1331: undefined reference to `drm_output_fini_egl'
| /.../libweston/backend-drm/kms.c:1332: undefined reference to `drm_output_init_egl'
```

It seems the code that is needed to fix this needs to be included by
enabling the build flag imxgpu. This is a bit confusing since i.MX 93
does not have a GPU.

The imxgpu config flag needs to be reworked, but for now simply enable
it for i.MX 93.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>